Finally fix the null bytes and duplication issue
`.insert` on a Vec **shifts** all elements after the index to the right.
What I wanted to do, is replace the value at the index.

Thats why the end result had duplicated data when the results streamed in a
non ascending order. And thats were the null bytes came from, especially
in the 0, 1, 2 case.